+++ packages/core/src/types/middleware.ts
@@ -6,7 +6,7 @@ export type DefaultTags = JSONObject;
 export interface BaseBillingMiddlewareOptions<
   TTags extends JSONObject = DefaultTags,
 > {
-  destinations: Destination<TTags>[];
+  destinations?: Destination<TTags>[];
   defaultTags?: TTags;
   waitUntil?: (promise: Promise<unknown>) => void;
   onError?: (error: unknown) => void;